Added KGModal and Toast Views.
KGModal provides flexible modal window functionality. This is used in
the help section of Coloring Book. When a user presses the help button
a modal window pops up with instructions on how to use the app.

The Toast View is a temporary notification that pops up on the screen
without being a nuisance. It is used when the user should know
something, but the notice is not critical.

Signed-off-by: Byron Sanchez <bfs5030@gmail.com>